Removes LOGGER.isDebugEnabled checks.
While there is nothing wrong with the application of the checks, and
they have some value when determining the debug string takes
non-neglectable time, the preference is to keep inline with the somewhat
shorter way logging statements are used throughout the project without
the check. If this turns out to be a problematic issue which can be
resolved by adding them back, of course we will do so.