added function in installation service to get and store scaler and unit from a E-meter
Implemented as just a function, not sending a response, meant to be called from installation process
or on demand or both. Exceptions are just thrown, functions that do send a response should handle these.

This function is just for E-meters, whereas the commandexecutor is for mbus devices as well. Mbus devices administration not yet in place.