Adds with-list capability check to smart meter message handling.
For devices unable to respond to get requests with-list, multiple get
requests will be issued. For this the command executor interface is
extended with the device as an input parameter. The device can then be
checked for capabilities when jDLMS calls are prepared.

Date time values with unspecified parts were observed as a result to a
call to one of the meters not supporting with-list requests. These are
given some default year/month/day for now, but this may need to be fixed
in some better way.