Sets default values for properties not in tariff schedule.
Next to setting defaults, the code for disabling the old schedule
entries has been optimized (enabled equals false is only written to the
device when the schedule entry was enabled to begin with).

Fixes the poorly named and documented (copy paste) method for
determining the name of the property of a schedule entry.