Changes handling bundle exceptions to include all action responses.
If a retryable exception occurs handling a bundle of actions, and the
maximum number of retries has been reached, the exception is propagated
through the platform.
These changes preserve any separate action responses before the
exception occurred, and places the exception as response for all actions
that did not yet have a response.