ZS-106 Removes references to indexed logical devices.
The LogicalDevice enum no longer contains a value for each index of a
logical device (e.g. PV1, PV2), but only one value for the type.
Other code referring to indexed logical devices is duplicated, adding a
second implementation that takes a logical device and an index.

This should enable get data, set data and reporting services not to be
constrained to a fixed number of instances of each logical device type.