SLIM-850 Allows missing configuration for public or private key
At the web service client side only the public key will be
available to encrypt device keys for device requests with keys.
This allows the configuration of the private key path in such a
case to be empty.
When only decryption is necessary the public key path does not
need to be configured.