SLIM-908 Removes dbs.url properties
With the change to the HikariDatasource the DB URL properties
that were configured are obsolete, since the database name from
the configuration is used with the other URL parts instead of the
entire JDBC URL itself.

The ResponseNotifierImpl had some sort of mix between repository
use and hard core result sets with prepared statements that
required a database URL. Replaced this with auto-wired
repositories and cleaned the code up a bit.