SLIM-841 Fixes Find Organization failures
Failures occurred in the FindOrganizations feature
because the list of organizations retrieved had the
organizations in a different order than expected.
Because the order is not important, the checks have
been altered to check if the expected organization
is contained in the response list, instead of also
checking the exact position.