SLIM-747 Adds firmware identification
Adds a unique identification to firmware to enable
look-ups by this unique key. This is to be used
instead of the filename that is not to be considered
a unique identifier any more.

Updates checks on firmware update to use the
identification of the firmware instead of the list
of version numbers to determine which firmware
module versions should have ended up on the device.