SLIM-1431 Adds temporary patches waiting for fix from SLIM-1387
SLIM-1387 should fix setting the M-Bus User key with a freshly generated
new key instead of reading the key from the database.
This fix is needed for setting the encryption key based on channel as
well.

This commit contains a temporary patch - in order to get the tests to
work - awaiting the cleaner solution to be merged in.