SLIM-1431 Adds test for Set M-Bus User Key By Channel operation
Adds a Cucumber test verifying the way Set M-Bus User Key By Channel is
handled.

The DlmsDeviceBuilder should include MbusIdentificationNumber and
MbusManufacturerIdentification.

Creating a device from the DlmsDeviceSteps logic for how to handle
device types and deal with keys is more aligned with the type instead
of looking whether a gateway device is identified in the step data.
This was driven partly by the fact that the relation between the M-Bus
device and the gateway device shouldn't be explicit in the step data for
the scenarios following an alarm for an M-Bus device on a certain
channel.

Fixes security key builder in the Cucumber preparation steps to be
usable with M-Bus keys.
* The builder should only set the key of the type it is configured for.
* The builder should also be able to deal with specific overriden values
  for M-Bus keys and the LLS password.