SLIM-1387 Couples M-Bus device based on Short ID
Couple M-Bus device looked for primary address non zero as first check
if a device was coupled on a channel. For wireless M-Bus devices the
primary address is always zero. If a device is accidentally not coupled
according to OGP a couple M-Bus device command leads to writing the
Short ID on the next free channel. These changes fix this by focusing
on the M-Bus Short ID when coupling an M-Bus device instead.