SLIM-1475 Keeps track of invocation counter for HLS 5 connection
Changes the code in order to be able to set an invocation counter (IC)
at the start of an HLS 5 connection, and keep track of the IC to prevent
the IC from being reused with the same encryption key. The IC is part of
the initialization vector used with HLS 5 encryption.
The IV should start with the system title (ASCII bytes of three
manufacturer ID chars followed by 5 bytes for the manufacturer device
identification). In order to be able to use a proper system title we
need to know the manufacturer ID for the device.

* Adds invocation counter to SecurityKey so a counter per key can be
  maintained.
* Sets the system title and the frame counter (IC) on the jDLMS TCP
  connection builder for use in the device connection.
* Adds an InvocationCountingDlmsMessageListener to be used as DLMS
  message listener with HLS 5 communication (which is extended by the 
  LoggingDlmsMessageListener used when a device is in debug mode).
* Derives the device ID long value for the system title from the DLMS
  device identification (twelve digits after the 5 character Kema code).
* For now, uses the M-Bus Manufacturer ID as manufacturer ID for the
  e-meter as well (at least until we have a way to properly maintain 
  this property). This should be the first three characters of the
  logical device name (accessible through the P3 public client as the
  value at OBIS code 0-0:42.0.0.255).
* Introduces DlmsConnectionMessageProcessor as abstract base class for
  request/response processors that need to be able to setup HLS 5
  connections with DLMS message listener.
* Refactors message processors for device request, update firmware
  request, OSGP response and get firmware file response in order to use
  a shared implementation of DLMS connection setup and post-processing.