SLIM-1426 Sanitizes use of ObisCode as path parameter
Some strange manipulation of the ObisCode formatted as short OBIS string
in the path was going on. This separates the class ID from the ObisCode
and sets up a converter that verifies the request has a valid ObisCode
as parameter.
In order to prevent issues that might arise in some contexts with
special characters the safest string representation (asDecimalString) is
used throughout the code where a textual representation of the ObisCode
is needed.