SLIM-1426 Updates code to use JSON objects related to DLMS simulator
Updates Cucumber step implementation code to translate data from step
tables to JSON objects to compare with the results from getting values
from the DLMS attribute values resource behind the DLMS simulator.

Updates features with DLMS simulator values to include the specific type
of number in order to build JSON objects matching DLMS data objects of
particular types.

Fixes ObisCodes in Cucumber steps where the 0 at the end of the short
notation was incorrectly written as the 255 of the last part of the full
code that is left out in the short notation.