SLIM-1426 Sets alarm object value from given step
Changes the test scenario about alarms on a device, making sure the
alarm object on the device has some alarms before retrieving them.

Updates the code behind to use a list of a list of strings as
representation of the table with attribute values instead of a DataTable
so it is easier to reuse it when the attribute data are in the code
instead of in the feature file.