OC-330 Changes re-send notification strategy for response data
Re-sends notification for response data based on the elapsed time since
the response data was created.
This allows a fixed time-before in combination with the number of
notifications sent in the response data query, getting rid of potential
query results that need to be skipped in processing.

Shared code from services extending AbstractResendNotificationService
has been put in the abstract class, which is now responsible for
updating the response data. The sub-classes now only need to concern
themselves with the check for a valid notification type and the actual
sending of the notification (which are domain-specific).