ZS-312 Updates message processor to prevent retries from queue
The message processor called from the context of a message listener's
onMessage method causes retries by the re-delivery mechanism configured
on the queue that is listened on when certain exceptions are thrown from
the implementation (JMSException, RuntimeException or Error).

Exceptions from sending notifications are only logged, and not re-thrown
to avoid re-delivery of the response data that may already have been
retrieved by the application in the mean time.

Exceptions extracting the meta-data that lead to unrecoverable error
being logged are now also IllegalArgumentException and
NullPointerException, since these could occur creating the enum values
or in some situations while getting JMS properties.