SLIM-869 Aligns notification service changes between SmartMetering and Microgrids
Changes the notification service implementation with the Microgrids
tests to be more in-line with the improvements in SmartMetering.

Fixes an issue with the new notification service that did not work
properly after a time-out on a CompletableFuture. This left the queue
blocked on a take from the code running asynchronously.