SLIM-833 Makes sure custom wait is only used on next call
The SmartMeteringBaseClient for the platform web service calls has a
method allowing the wait period for a notification to be customized for
the next web service call from a scenario step.
The implementation however changed the wait period for all following
service calls for the client where it was set.
This makes sure that a custom wait is applied when it was set, but is
reset after use, allowing the configured wait time to be used for other
service calls.