SLIM-833 Only sets longer wait time when scenario is tagged as nightly
For the step getting the administrative status in such a way that it
returns an exception, some of the scenarios require a longer wait time.
This makes sure the long wait time is set if the scenario is tagged for
NightlyBuildOnly, and if not the generally configured wait time is used.