SLIM-833 Makes DlmsPushNotificationDecoder more robust for delayed bytes
If the pushed bytes for a DlmsPushNotification sent by a smart meter are
not received fast enough, it is possible that the decoder encounters
only part of the bytes it needs to process the data to get to the next
state.

When dealing with the equipment identifier an error could occur if the
comma byte was not included in the bytes received (end index of the
range to copy from a byte array would be -1).

These changes make sure enough bytes are read from the channel buffer to
be able to determine the equipment identifier and having read the comma.
For the ReplayingDecoder it is not an issue if not enough bytes are
available. It is designed to deal with this behind the scenes by using
an internal ReplayError.
Changes in this commit allow this replay mechanism to make sure enough
bytes are received for proper decoding instead of counting the bytes in
the DlmsPushNotificationDecoder itself.

The changed implementation does not build push notifications that can
not be handled if no bytes are available, because this is not needed.

Decoding the alarm bytes is done by reading them as integer instead of
determining the value one byte at a time.

Fixes some minor Sonar issues as well.