Fixes DLMS ConfigurationObject related tests
Cucumber tests for retrieving the configuration object could fail when
another test setting the configuration object sets unexpected values on
this object earlier in the test run.

Updates the SimulatorTriggerClient cleaning up and removing duplication.
Checks not only for HTTP status OK for success, but allows any status
from the success-family, returning null on 204-NO_CONTENT.
Adds methods to get or set a single attribute value instead of having to
get or set all values for an object with a certain OBIS code.

Updates the GetConfigurationObjectHelper class to make sure meaningful
errors are thrown in case of failures that may be expected, instead of
having NullPointerExceptions or ClassCastExceptions to examine further
for what actually happened.