Fixes DLMS ConfigurationObject related tests
Cucumber tests for retrieving the configuration object could fail when
another test setting the configuration object sets unexpected values on
this object earlier in the test run.

Updated BundledGetConfigurationObject feature. The background now sets
the configuration object for the simulator, so it is known what may be
expected to be returned from the calls through the platform. Added
HLS_5_ON_P_3_ENABLE set to true, because it does not make sense for that
flag to be false.

Updated SetConfigurationObject feature. The background now sets
the configuration object for the simulator, so it is known what may be
expected to be on the simulated device after the set actions from the
scenarios.
Adds verification that the simulated configuration object is updated as
expected from the executed set requests.

Added methods for getting or setting a single attribute value to the
DeviceSimulatorSteps to avoid retreiving/creating objects for all
attribute values of a COSEM interface object, when only a single one is
important to the test at hand.
Refactored DeviceSimulatorSteps, which had too many reasons to change.
Code for specific simulated objects is moved to
SimulatedAlarmObjectSteps and SimulatedConfigurationObjectSteps. Code
for creating the JSON objects required for the API calls is moved to
JsonObjectCreator.
Refactored to remove duplication in the error handling around web client
calls.

Removed some redundant logging and asserting from
SetConfigurationObject.