SLIM-1134 Fixes replace keys scenario without functional exception
The scenario to replace keys with generated ones was copied from the
FunctionalExceptionsEncryptionKeys feature to the
SmartMeteringConfiguration feature because it does not involve a
functional exception.
The given steps to make sure the meter exists were still in the scenario
where they are part of the background of the configuration feature.
This removes the devices from the scenario to prevent JPA exceptions
trying to add them to the database two times.