OC-442 Adds abstract base class for RTU read command factories
Moves implementation of instance methods to
AbstractIec61850RtuReadCommandFactory. Classes extending this abstract
class need to define attributes using a filter ID, and need to configure
start and end (inclusive) values for the ID in filtered attributes. The
commands with the attributes are set up in the base class.

Some classes had a static instance field that is lazily initialized
(for instance Iec61850CombinedLoadCommandFactory), the idiom used is
changed to the one recommended in Effective Java 3rd edition (using a
static holder class for the instance field).