OC-380 Updates code based on SonarQube reports and review comments
Replaces the NotificationService definition of sendNotification that has
too many parameters by one where many of the parameters are wrapped in a
NotificationWebServiceLookupKey and a GenericNotification.

Renames DataBasedNotificationService to DefaultNotificationService, to
indicate that this is the default implementation (next to the
NotificationServiceBlackHole no-op implementation).

Makes templateFactory required for the DefaultNotificationService.
Since it implements NotificationService, the
NotificationServiceBlackHole should be configured if no notifications
should be sent, instead of leaving templateFactory null.

Renames WebServiceConfigurationDataId to NotificationWebServiceLookupKey
and makes it a top level class, to better fit with its usage not only as
database ID.

Adds constructors with arguments to GenericSendNotificationRequest and
GenericNotification to avoid the need for multiple setter calls and
separate construction methods in places they are created.

Suppresses the warning about an intentionally ignored exception in the
DataBasedWebServiceTemplateFactory method getTemplate.