OC-380 Updates NotificationService for Smart Metering
Makes the Smart Metering notification service suitable for multiple
endpoints.

Updates the NotificationService to allow specific implementations to
override the target URI based on anything from the notification inputs.

Changes NotificationServiceWs to
CorrelationUidTargetedNotificationService, extending
DefaultNotificationService to add a possible override for the target URI
if a response URL was saved with the correlation UID.