OC-61 Uses new enum MessageType as key for the message processor map
Adds a new enum MessageType to replace other enums from platform and
protocol adapters. This MessageType is to be used as source for the
JMS type that identifies the device function through the layers of OSGP.
By using this single enum, it should no longer be necessary to duplicate
the same enum names (and sometimes even ordinal position) accross
several enums spread throughout the code.

Because the single enum is used, BaseMessageProcessorMap can provide a
single implementation for getMessageProcessor, and the different sub-
classes don't need to include an implementation that is nearly the same.