OC-61 Replaces device function and message type enums with MessageType
Changes references to the message type which is included as JMS type for
the queues between the layers to be based on a single enum (MessageType)
to prevent duplication of enum names (and sometimes even ordinal
positions) that are required to match up in different enums.

Removes getMessageProcessor from the message processor map
implementations as it is defined in the BaseMessageProcessorMap now.

Tries to consistently use name instead of toString when the MessageType
enum needs to be put into a String to be used later-on with valueOf.

Removes the following types (no longer used):
 * AdminRequestMessageType
 * CommonRequestMessageType
 * DistributionAutomationRequestMessageType
 * MicrogridsRequestMessageType
 * PublicLightinhRequestMessageType
 * SmartMeteringRequestMessageType
 * TariffSwitchingRequestMessageType

Addresses some SonarQube warnings:
 * removes unnecessary Object array with LOGGER
 * removes throws clause for exception that cannot be thrown
 * fixes log statements where the number of arguments did not match
 * removes (effectively) unused local variables
 * removes unused autowired fields
 * removes unnecessary toString on Strings
 * uses placeholder instead of string concatenation with logger

Removes unnecessary formatter on/off comments.