OC-61 Replaces DeviceRequestMessageType with general MessageType
Uses new enum MessageType to reference the JMS type as it is put into
and read from JMS metadata, instead of DeviceRequestMessageType which
duplicates a lot of the values.

Removes getMessageProcessor from the message processor map
implementations as it is defined in the BaseMessageProcessorMap now.

Removes the following types (no longer used):
 * DeviceRequestMessage
 * DeviceRequestMessageType