FLEX-5239 ~ Fixes OSLP Configuration scenario
The OSLP mock server used to be configured with null for the interval
type when the matching cell in the step data was empty.
This change makes sure the default interval type is only used if there
is no step data, and null is used if there is empty interval type step
data.