FLEX-5239 ~ Upgrade to Cucumber 5 (#153)
* FLEX-5239 ~ Starts OSGP upgrade to Cucumber 5

Refactors packages of Cucumber classes to the newer API. The core part
is considered internal and to be used through the java and junit
projects.

Add strict equals true to the AcceptanceTests classes as the non-strict
mode is deprecated, and will be removed in a future version. This
prevents warnings in the output logs.

Updates settings handling in the steps, to get rid of issues because of
empty cells defaulting to null instead of the empty string as before.

Updates step method annotations where the step definition text with the
step methods is not wrapped between "^" and "$" and the step has
parameters. These are now considered Cucumber expressions instead of
regular expressions. For Cucumber expressions a different syntax is
needed to match the arguments.

Updates the Configurer to no longer implement TypeRegistryConfigurer.
The configurers registers some default step data transformers without
implementing an interface of which only one implementation is allowed on
the glue path. This should offer flexibility in adding custom
transformer functionality with other step classes as needed, without
having to configure this in a single class.

Adds the junit-vintage-engine to the POMs of projects with Cucumber
tests. The current setup of Cucumber depends on JUnit 4.

Uncommented and at-skipped a feature in
TariffSwitchingResponseDataCleanupJob.feature in
cucumber-tests-platform-publiclighting. A feature file without any
contents leads to errors when running the Cucumber tests.

* FLEX-5239 ~ Fixes Cucumber expression steps smart metering

Fixes some steps that are Cucumber expressions instead of regular
expressions in the smart metering SimulatedConfigurationObjectSteps.

Bumps the Cucumber version to the latest release (5.1.2).

* FLEX-5239 ~ Replaces regex by Cucumber expression

* FLEX-5239 ~ Updates triggering the Cucumber tests

Cucumber test execution using command line argument cucumber.options has
been deprecated. These changes should make sure the old options that
were used are replaced by cucumber.execution.strict and
cucumber.filter.tags.

For cucumber.filter.tags, the parsing of tags from the pull request
description, used to restrict the subset of the tests to be executed, is
updated, to no longer result in multiple --tag sections, but instead to
make an expression using and and or.

* FLEX-5239 ~ Fixes missing escaped backslash in pipeline shell script

* FLEX-5239 ~ Fixes OSLP Configuration scenario

The OSLP mock server used to be configured with null for the interval
type when the matching cell in the step data was empty.
This change makes sure the default interval type is only used if there
is no step data, and null is used if there is empty interval type step
data.