OC-533 Makes build fail on error running tests
When an error is introduced in the Spring or glue context of Cucumber
tests an error is thrown preventing further tests to be executed.
This happens because the Cucumber feature runner makes the JUnit runner
abort the tests throwing a StoppedByUserException.
In such cases the Jenkins console log contains the text "End. Retval =
1", but does not contain any information on failed scenarios.
The LogParserPublisher Jenkins plugin which is activated in order to
check the log for issues with the build has configuration in a file
named console-test-rules marking this situation as an error.
By adding parameter failBuildOnError with value true, the build will be
failed when the LogParserPublisher determines an error has occurred.