OC-533 ~ Modifies test result when all failures are in the mechanism
Failures in the test mechanism are related to errors during the test run
that could not be related to test cases. Modify the test result to avoid
failing the build in case all failures are test mechanism related and
some tests have succeeded.

Log test mechanism related failures with their exception traces so they
can be examined from the log to see if the test setup can be changed to
prevent them from occurring.