OC-533 Makes build fail on error running tests (#161)
* OC-533 Makes build fail on error running tests

When an error is introduced in the Spring or glue context of Cucumber
tests an error is thrown preventing further tests to be executed.
This happens because the Cucumber feature runner makes the JUnit runner
abort the tests throwing a StoppedByUserException.
In such cases the Jenkins console log contains the text "End. Retval =
1", but does not contain any information on failed scenarios.
The LogParserPublisher Jenkins plugin which is activated in order to
check the log for issues with the build has configuration in a file
named console-test-rules marking this situation as an error.
By adding parameter failBuildOnError with value true, the build will be
failed when the LogParserPublisher determines an error has occurred.

* OC-533 ~ Adds test result failure details logging

Because there are some builds where the Cucumber test results are all
successful, yet the all-over result has failed because a failure is
present in the JUnit run results, detailed failure logging is added to
investigate what can be the cause of this.

* OC-533 ~ Modifies test result when all failures are in the mechanism

Failures in the test mechanism are related to errors during the test run
that could not be related to test cases. Modify the test result to avoid
failing the build in case all failures are test mechanism related and
some tests have succeeded.

Log test mechanism related failures with their exception traces so they
can be examined from the log to see if the test setup can be changed to
prevent them from occurring.