FLEX-5274 ~ Fixes issues in Cucumber set up IEC 60870
Removes version properties for Cucumber and Spring Test from the
protocol adapter IEC60870 POM to align with versions from the super POM.
Adds a dependency to the JUnit vintage engine so the TestRunnerIT will
be executed to trigger the Cucumber component tests.

Fixes copy paste mistakes in the Cucumber tests for platform
distribution automation. Removes reference to features not present in
this project from the CucumberOptions in AcceptanceTests.
Changes the mainClass in the POM for the jar with dependencies to the
App that belongs with the distribution automation project and which is
present in the resulting jar, so that the jar can be executed.

Adapts the Jenkins pipelines to include Distribution Automation tests.