FLEX-5274 ~ Fixes indentation of dependency in POM
Adjusts the indentation of an added dependency to the JUnit vintage
engine to match formatting of the surrounding POM elements.

Switches the branch of the release repo in the pull request build
pipeline to use a test server with some additional distribution
automation components. The one from the master branch appears to be
missing at least the protocol IEC60870 related stuff.