FLEX-5274 ~ Pefixes feature description with DistributionAutomation
The results in the Cucumber reports on the build server are grouped by
natural ordering of the feature description. Update the feature
description so that distribution automation Cucumber tests stand out
among those from other streams and will be grouped together if more
features are added adhering to the convention.