FLEX-5279 ~ Modifies openiec61850 usage to try and fix simulation
Sets the RTU simulator server model as the model copy from the parsed
server.

Puts back the client association in the methods from the device
connection service, in order to explicitly set the server model on it
when the model is read from a file instead of from the connected device.
Explicitly sets the server model on the client association when parsed
from a file in the Iec61850Client.