OC-557 ~ Prevents long running database transaction.
The function `sendEventNotificationCommand()` was annotated with
`@Transactional`. This made sure the `OslpLogItem` could be saved. But
this function also creates a TCP connection to actually send an event
from the simulator to the protocol adapter. In case this TCP connection
hangs/errors/times-out the database transaction will be blocked during
that time.

The new class `OslpLogService` makes sure that the database transaction
required for saving an `OslpLogItem` is small and fast.