OC-526 ~ Adds dependency analysis reporting
Configures the aggregate report of the OWASP dependency-check-maven
plugin to analyze dependencies with potential security vulnerabilities.

Configures reports about dependency and plugin updates by the
versions-maven-plugin to get insight into dependencies and plugins with
newer versions than used.

Configures a system dependency for tools.jar (com.sun:tools:1.8.0) that
caused errors when missing during the dependency analysis. Takes into
account that this jar may not be present for some Java installations on
Mac OS and uses a suggested alternative (not tested).

Replaces the FindBugs plugin, which is no longer maintained, by its
successor the SpotBugs plugin. Sets the version in the POM (which was
missing for the FindBugs plugin causing errors during dependency
analysis in the build).

Replaces PMD and Checkstyle versions by the versions of the respective
maven plugins (the version of the plugin is different from the version
of the tool, which caused errors during dependency analysis since the
plugins did not exist for the versions of the tools in the POMs).

Cleans up plugin management configuration of the PMD, Checkstyle and
SpotBugs plugins.