OC-526 ~ Adds Jenkins pipeline for dependency reports
Adds a Jenkins pipeline to run the maven site command and publish HTML
reports that are configured for dependency checks and updates.

Configures rules for the maven versions plugin to prevent alpha, beta,
milestone and release candidate versions to be reported as newer
versions that are available for dependencies or plugins.

Configures the maven site plugin to the latest version currently
available.