OC-355 ~ Schedules processing Scheduled Tasks with Quartz
Processing scheduled tasks was scheduled using another mechanism than
Quartz. All actions sensitive to concurrent execution in a scaled
environment with potentially multiple instances of the code running is
scheduled with Quartz to effectively disallow concurrent execution.

Refactors SchedulingConfig to ScheduledTaskExecutorJobConfig wiring the
Quartz based OsgpScheduler to create and schedule the job.

Splits the ScheduledTaskScheduler in ScheduledTaskExecutorJob
(implementing the Quartz Job interface) and ScheduledTaskExecutorService
(implementing processing and execution of scheduled tasks).

Adds OsgpSchedulerConfig to make an OsgpScheduler available to the code
in osgp-core.

Updates the ScheduledTaskSchedulerTest and renames it to
ScheduledTaskExecutorServiceTest to follow the other refactorings and
because this code is not about scheduling of the scheduled tasks, but
about processing/executing them.

Updates osgp-core.properties replacing the now obsolete properties
scheduling.task.scheduler.pool.size and
scheduling.task.scheduler.thread.name.prefix by
quartz.scheduler.thread.count.
Alters the value of scheduling.scheduled.tasks.cron.expression, because
the existing value leads to a ParseException with Quartz (one of the
day-of-week and day-of-month fields must be a question mark with Quartz
cron expressions).