FLEX-5289 ~ Configures javax.annotation-api in super POM (#246)
Adds the dependency to javax.annotation-api in the dependency
management section of the super POM.
In many projects the dependency is transitively present, and
mostly used in for example the javax.annotation.PostConstruct
annotation.
If it is used and not declared, maven dependency analysis will
warn about a used undeclared dependency. Having the dependency
configured in the super POM simplifies fixing the maven warning
without duplication of version metadata.