FLEX-4693 Replaces FirmwareFile constructors with Builder
Based on a review comment indicating there were too many constructors
with growing and confusing lists of arguments, a Builder is added to
FirmwareFile replacing the constructors.

Device models and firmware modules with versions were kept outside the
builder, because of comments in FirmwareManagementService with
addFirmware that the FirmwareFile needs to be saved before relating
other existing entities to prevent issues with Hibernate.