FLEX-5297 ~ Sends end of initialization ASDU on first connection
Updates the Iec60870ServerEventListener, so it can be created in a way
that makes sure it sends an end of initialization ASDU when the first
connection to the server is created.
The current implementation assumes a couple of things, that may need to
be modified with later insights (for instance the information object
address to be used is set to 0, and the IeCauseOfInitialization is
created with values 0 and false).

The Iec60870SimulatorConfig is extended with a boolean value
(sendEndOfInitialization) indicating whether the end of initialization
ASDU is sent when a first connection is established or not. This value
is used when creating an Iec60870Server as option with the server event
listener.

A client ASDU handler (EndOfInitializationAsduHandler) is added, mainly
to prevent errors about unknown ASDU types in the logging. For now its
implementation only has some informational log output.