FLEX-5395 ~ Adds configuration of IP addresses that may be duplicated
Upon device registration, the IP address of the registered device is
removed in the database for other devices where it is configured (with
the exception of the loopback address - localhost / 127.0.0.1).

These changes allow to either configure that such clean up should never
be performed (allowing multiple devices to share any network address),
or to configure a comma separated list of IP addresses or IP address
ranges that may be shared (for instance
"192.168.0.1,192.168.1.12-192.168.2.103" to allow sharing of 192.168.0.1
or any IP address from 192.168.1.12 to 192.168.2.103 inclusive).