OC-595 ~ Adds logging of Kafka consumer records
Adds KafkaLogger for use with consumer configuration on Kafka topics to
be able to log data and metadata of messages that are posted on a topic.

Adds osgp-adapter-kafka-logging as deployable project in servlet
containers with configuration of consumers for currently known topics
in the GXF code base.

The configurations of the different logging consumers are subject to a
Spring conditional, following a boolean property (for instance
peakshaving.kafka.logging.enabled). Default this enabled property is set
to false, which should prevent warning or error logging if a Kafka
broker is not available for certain topics in a monitored environment.

The consumed messages for the configured topics will be logged to a file
specified in the logback configuration with an appender named MESSAGES.
This appender leaves out most of the default stuff that is normally
added as context with logged data, leaving only the date and time of
logging, and the message that is logged.